project/requirements.txt
torch==2.0.0
torchvision==0.15.0
numpy==1.24.0
pandas==2.0.0
matplotlib==3.7.0
seaborn==0.12.0
scikit-learn==1.2.0
Pillow==9.5.0
shap==0.42.0
lime==0.2.0
opencv-python==4.7.0
jupyter==1.0.0
ipykernel==6.22.0
